requirements.txt
speechrecognition
pyttsx3
groq
pillow
pytesseract
datasets
torch
transformers
python-dotenv
soundfile
sounddevice
flask
requests
beautifulsoup4
keyboard